Take fallback value into account
The patch modifies the logic of BaseInputFilter to ensure that if an
input supports fallback values (is an instanceof Input), and has
fallback value, then that fallback value will be considered when
determining validity of the input. In other words, if the value is
missing from the data set, the fallback value is assumed.

It also adds a test to ensure that if the value *is* passed but empty, the
fallback value is used for that input.