Remove test about Input setters permutation
These tests asserts the state of Input is always the same not matter the order when invoke the setters.

Sometime ago there was specific tweaks where setterA change the default value for propertieB. This feature was removed so this tests are verifyng "Feature X" not longer exists.

PD: setRequired is already verified by testRequiredFlagIsMutable