Do not allow setting invalid collections or collection items
Unlike the base input filter, `CollectionInputFilter::setData()` has
always allowed *any* data in. It only raises error conditions in one
situation: when validating, if the data is not an array or traversable.

Otherwise, as it iterates through the data, if any item is not an array,
it casts it to... an empty array.

This patch updates the behavior of `setData()` to do the following:

- If the `$data` argument is not an array or `Traversable`, it raises an
  `InvalidArgumentException`.
- If any item in `$data` is not an array or `Traversable`, it raises an
  `InvalidArgumentException`.

Unfortunately, this change now introduces six errors in the test suite.
From my cursory inspection, each of these are malformed tests, in my
opinion, as they represent invalid data in the first place.

A later commit will take care of those issues.