Corrected data provider
The data provider `countVsDataProvider()` was incorrectly providing
collections. `$data1` needed to be an array containing a single array,
and `$data2` needed to be an array containing two arrays in order to
properly setup the input filter for the given expectations.

The provider values worked previously due to the fact that `isValid()`
was not being called, which would have made the counts invalid as the
data was invalid.