Add zend-servicemanager as a required dependency
After being bitten by the bug in #119, where I couldn't use
zend-inputfilter without zend-servicemanager, I'm proposing this PR,
which makes zend-servicemanager a required dependency. I can't claim to
understand all the reasoning for not doing so to date, but based on this
experience, it makes sense to make the change.