Ensures that `null` data passed to an input filter is cast to an empty array
Per #159, `BaseInputFilter` works inconsistently when a `null` value is
passed to `setData()`, interpreting this as an invalid argument type
versus an empty data set. This is particularly problematic with nested
sets sent via an API, as they may be nullable. In such cases, they
should be treated the same as if an empty array were provided.