Allow setting a failure message for required Collections
This patch modifies the `CollectionInputFilter` to allow injecting a
`Zend\Validator\NotEmpty` instance. When the collection is marked
"required", but receives an empty data set, it will use that validator
to provide a validation failure message. (If no validator was set, it
lazy-loads one.)

When using factory-based collections, you may now also provide a
`required_message` key as a sibling to `required`; when set, this
message will be used in place of the default `NotEmpty` validation error
message.